+++ CHANGES.rst
@@ -14,6 +14,7 @@ Unreleased
 * Use indexes for :meth:`~Reader.get_entries()` (recent order);
   should make calls 10-30% faster.
   (:issue:`134`)
+* Allow sorting :meth:`~Reader.search_entries` results randomly. (:issue:`200`)
 
 
 Version 1.9